Add Swich Games dropdown to navbar, and adds favorites
This commit allows users to switch between games. The dropdown contains
a list of a user's favorite games. When a game is selected, the User’s
game_setting_id attribute is updated, and returns the user back to
whichever page initiated the request.

All of the user’s favorites(currently strategies and games) are handled
by the acts_as_favoritor gem.